Implement new strategy for inheritance with poll items + Create open text (#75)
* Change IDs from int to long

* Delete unused files regarding Live-Poll answers

* Change more ids from int to long

* Add/Adjust entities, DTOs and repos for poll items

* Implement new inheritance strategy for poll items

* Implement use case "Create open text item"

* Update postman collection with create open text

+++ src/main/kotlin/de/livepoll/api/entity/db/BlockedToken.kt
@@ -1,6 +1,6 @@
 package de.livepoll.api.entity.db
 
-import java.util.Date
+import java.util.*
 import javax.persistence.*
 
 @Entity
@@ -9,7 +9,7 @@ data class BlockedToken(
         @Id
         @Column(nullable = false)
         @GeneratedValue(strategy = GenerationType.IDENTITY)
-        val id: Int,
+        val id: Long,
 
         @Column(nullable = false)
         val token: String,
