+++ src/main/kotlin/de/livepoll/api/entity/dto/PollDtoOut.kt
@@ -1,10 +1,9 @@
 package de.livepoll.api.entity.dto
 
-import de.livepoll.api.entity.db.PollItem
-import java.util.Date
+import java.util.*
 
 data class PollDtoOut(
-        val id: Int,
+        val id: Long,
         val name: String,
         val startDate: Date,
         val endDate: Date,
